feat helm 설정을 위한 dockerfile 수정
- ENTRYPOINT에 jar 실행 시점을 cmd로 분리
- -> springboot helm configmap.yaml의 application.yml args로 반영하기 위함
