+++ webtoon-search-back/webtoon-search/docker/Dockerfile-spring
@@ -50,4 +50,5 @@ RUN mkdir -p /home/appuser/logs && \
 
 # 애플리케이션 실행 // ENTRYPOINT - 무조건적인 실행, CMD - 실행 시의 옵션에 따라 변경 가능
 # -Djava.security.egd 옵션은 암호화 작업의 엔트로피 소스를 지정하여 시작 시간 개선
-ENTRYPOINT ["java", "-Djava.security.egd=file:/dev/./urandom", "-jar", "/app/app.jar"]
+ENTRYPOINT ["java", "-Djava.security.egd=file:/dev/./urandom"]
+CMD ["-jar", "/app/app.jar"]
